Change repo into a basic Flask Hello World
Removed the blog code, as it posed problems that are not project
specific. For now a simple Hello World will suffice.

Added .dockerignore, so that Dockerfile COPY directive wont copy all
the loose files in repo onto the docker image.

Created a Dockerfile that build a working image based on that code.